Move around BSP-related files, re-structure task logger
