+++ Sources/SourceKitBazelBSP/SharedUtils/SplitLogMessage.swift
